+++ data-retrieval/taxonomy-fetcher/src/impls/ncbi/taxonomy_builder.rs
@@ -23,13 +23,13 @@ pub struct NCBITaxonomyBuilder {
     /// Set the directory where the taxonomy is stored.
     directory: Option<std::path::PathBuf>,
     /// Root of the taxonomy.
-    root_position: Option<u32>,
+    root_position: Option<usize>,
     /// Taxon entries.
     taxon_entries: Vec<NCBITaxonEntry>,
     /// Hashmap from taxon name to position in taxon entries.
-    name_to_position: std::collections::HashMap<String, u32>,
+    name_to_position: std::collections::HashMap<String, usize>,
     /// Hashmap from taxon identifier to position in taxon entries.
-    id_to_position: std::collections::HashMap<u32, u32>,
+    id_to_position: std::collections::HashMap<u32, usize>,
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
@@ -195,7 +195,7 @@ impl TaxonomyBuilder for NCBITaxonomyBuilder {
         &self,
         id: &<Self::TaxonEntry as crate::traits::TaxonEntry>::Id,
     ) -> Option<&Self::TaxonEntry> {
-        self.id_to_position.get(id).map(|&pos| &self.taxon_entries[pos as usize])
+        self.id_to_position.get(id).map(|&pos| &self.taxon_entries[pos])
     }
 
     async fn build(
@@ -279,13 +279,13 @@ impl TaxonomyBuilder for NCBITaxonomyBuilder {
                 .set_rank(record.rank)?
                 .build(&self)?;
 
-            self.id_to_position.insert(taxon_entry.id, self.taxon_entries.len() as u32);
-            self.name_to_position.insert(taxon_entry.name.clone(), self.taxon_entries.len() as u32);
+            self.id_to_position.insert(taxon_entry.id, self.taxon_entries.len());
+            self.name_to_position.insert(taxon_entry.name.clone(), self.taxon_entries.len());
             if record.parent_tax_id == record.tax_id {
                 if self.root_position.is_some() {
                     return Err(crate::errors::TaxonomyBuilderError::MultipleRoots);
                 }
-                self.root_position = Some(self.taxon_entries.len() as u32);
+                self.root_position = Some(self.taxon_entries.len());
             }
             self.taxon_entries.push(taxon_entry);
         }
