+++ data-retrieval/taxonomy-fetcher/src/impls/open_tree_of_life/taxonomy.rs
@@ -46,7 +46,7 @@ impl Taxonomy for OpenTreeOfLifeTaxonomy {
 
     fn root(&self) -> Self::Taxon<'_> {
         OpenTreeOfLifeTaxon {
-            taxon_entry: &self.taxon_entries[self.root_position as usize],
+            taxon_entry: &self.taxon_entries[self.root_position],
             taxonomy: self,
         }
     }
